requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-cloud-firestore>=2.19.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
python-multipart==0.0.9
pydantic==2.9.2
